play-requirements.txt
pytorch
python-chess
